+++ tomcat/src/main/java/org/apache/coyote/http11/RequestLineParser.java
@@ -1,5 +1,11 @@
 package org.apache.coyote.http11;
 
+import org.apache.coyote.http11.model.HttpHeaders;
+import org.apache.coyote.http11.constants.HttpMethod;
+import org.apache.coyote.http11.model.HttpRequest;
+import org.apache.coyote.http11.model.RequestBody;
+import org.apache.coyote.http11.model.RequestTarget;
+
 public final class RequestLineParser {
 
     public static final String SPACE = " ";
