+++ tomcat/src/main/java/org/apache/coyote/http11/constants/HttpMethod.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.constants;
 
 import java.util.Collections;
 import java.util.Map;
