+++ tomcat/src/main/java/org/apache/coyote/http11/constants/HttpStatusCode.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.constants;
 
 public enum HttpStatusCode {
     SUCCESS(200),
