+++ tomcat/src/main/java/org/apache/coyote/http11/model/QueryParamsMap.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.model;
 
 import org.apache.commons.lang3.ObjectUtils;
