+++ tomcat/src/main/java/org/apache/coyote/http11/model/RequestTarget.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.model;
 
 public record RequestTarget(
         String path,
