+++ tomcat/src/test/java/org/apache/coyote/http11/HttpPathTest.java
@@ -1,5 +1,6 @@
 package org.apache.coyote.http11;
 
+import org.apache.coyote.http11.model.RequestTarget;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
