+++ tomcat/src/test/java/org/apache/coyote/http11/HttpQueryParamsTest.java
@@ -1,5 +1,6 @@
 package org.apache.coyote.http11;
 
+import org.apache.coyote.http11.model.QueryParamsMap;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
